Split radiomics feature families into dedicated modules and remove legacy definitions
