+++ zrad/radiomics/__init__.py
@@ -1,1 +1,1 @@
-from .radiomics import Radiomics
+from .extractor import Radiomics
